Rename to appConfig.js
