+++ lib/appConfig.js
