+++ lib/webpackConfigurator.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const config = require('./config');
+const config = require('./appConfig');
 const deepAssign = require('deep-assign');
 const webpack = require('webpack');
